seal template controller
