+++ test/docfx.Test/lib/TemplateTest.cs
